w2/db_query/backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.28.0"]
build-backend = "hatchling.build"

[project]
name = "db-query-backend"
version = "0.1.0"
description = "FastAPI backend for the DB query tool"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.128.7",
  "openai>=2.18.0",
  "psycopg2-binary>=2.9.11",
  "pydantic>=2.12.5",
  "sqlglot>=28.10.1",
  "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
  "mypy>=1.19.1",
  "pytest>=9.0.2",
  "ruff>=0.15.0",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
warn_return_any = true
show_error_codes = true
pretty = true

[[tool.mypy.overrides]]
module = ["psycopg2", "psycopg2.*"]
ignore_missing_imports = true

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
